requirements.txt
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
scipy==1.14.1

torch==2.4.1
# tensorflow-cpu==2.17.0

transformers==4.45.2
datasets==3.0.2
accelerate==1.0.0
sentencepiece==0.2.0

tqdm==4.66.5
matplotlib==3.9.2
seaborn==0.13.2
jupyter==1.1.1

sqlalchemy==2.0.36
psycopg2-binary==2.9.9
sqlite-utils==3.37

yamlargparse==1.31.1
python-dotenv==1.0.1

black==24.8.0
ruff==0.6.8

# Social data collection
praw==7.7.1
snscrape==0.7.0.20230622
beautifulsoup4==4.12.3
lxml==5.3.0
librosa==0.10.2.post1
soundfile==0.12.1

# Advanced NLP and analysis
nltk==3.8.1
emoji==2.8.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2

# Mood detection and ML models
torch==2.4.1
transformers==4.45.2
datasets==3.0.2
accelerate==1.0.0
sentencepiece==0.2.0
tf-keras==2.15.0

# Recommender system dependencies
implicit==0.7.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0